+++ src/app/shared/models/newTaskEntry.model.ts
@@ -1,0 +1,9 @@
+export interface NewEntry {
+  userId: number;
+  projectTaskId: number;
+  durationInMin: number;
+  note: string;
+  day: number;
+  month: number;
+  year: number;
+}
