+++ src/app/shared/models/task.model.ts
@@ -1,0 +1,5 @@
+export interface Task {
+  taskID: number;
+  name: string;
+  billable: boolean;
+}
