+++ src/app/shared/models/userProject.model.ts
@@ -1,0 +1,9 @@
+import { Task } from './task.model';
+
+export interface UserProject {
+  projectID: number;
+  projectName: string;
+  clientName: string;
+  projectCode: string;
+  tasks: Task[];
+}
