+++ src/app/time-sheet/components/time-entry-dialogue/time-entry-dialogue.component.html
@@ -3,11 +3,10 @@
   <p class="subtitle"> {{currentDate.format("dddd, D MMM")}}</p>
 </div>
 
-<form [formGroup]="TimeEntryForm" #form="ngForm" (ngSubmit)="onSubmit()">
+<form [formGroup]="TimeEntryForm" (ngSubmit)="onSubmit()">
 
   <div class="content">
     <mat-dialog-content>
-
       <mat-form-field>
         <mat-select formControlName="project" placeholder="Project" class="dropdown-input" required>
           <mat-option *ngFor="let p of userProjects" [value]="p" (onSelectionChange)="toggleTaskDropdown()">{{p.projectName}}</mat-option>
@@ -39,7 +38,7 @@
   <div class="actions">
     <mat-dialog-actions>
       <button mat-button cdkFocusInitial color="primary" type="submit" [disabled]="TimeEntryForm.untouched">Add Entry</button>
-      <button mat-button mat-dialog-close (click)="onNoClick()" color="warn" type="button">Cancel</button>
+      <button mat-button mat-dialog-close (click)="closeDialog()" color="warn" type="button">Cancel</button>
     </mat-dialog-actions>
   </div>
 </form>
